+++ src/sysinfo.rs
@@ -72,6 +72,7 @@ mod tests {
             location_ip_address,
             location_sqlite,
             log_level: tracing::Level::INFO,
+            rainbow: None,
             sql_threads: 1,
             start_time: SystemTime::now(),
             timezone: EnvTimeZone::new("America/New_York"),